Combine ExternalOverlap with External
Keep all the existing doc examples and unit tests. Implement early exit
checks directly in External. Add site_energy_initial and
is_only_infinite_or_zero to allow skipping the evaluation of the initial
energy for hard overlap potentials.
