+++ hoomd-interaction/src/external_type.rs
@@ -19,12 +19,9 @@ use hoomd_microstate::{Body, Microstate, Transform, boundary::Wrap, property::Po
 ///
 /// [`external`]: crate::external
 ///
-/// Use [`ExternalOverlap`] instead of [`External`] for purely hard interactions.
+/// # Examples
 ///
-/// [`ExternalOverlap`]: crate::ExternalOverlap
-///
-/// # Example
-///
+/// A linear external potential:
 /// ```
 /// use hoomd_interaction::{External, TotalEnergy, external::Linear};
 /// use hoomd_microstate::{Body, Microstate, property::Point};
@@ -48,6 +45,43 @@ use hoomd_microstate::{Body, Microstate, Transform, boundary::Wrap, property::Po
 /// # Ok(())
 /// # }
 /// ```
+///
+/// Infinite interaction with a wall:
+/// ```
+/// use hoomd_interaction::{External, SiteEnergy, TotalEnergy};
+/// use hoomd_microstate::{Body, Microstate, property::Point};
+/// use hoomd_vector::Cartesian;
+///
+/// struct Wall;
+///
+/// impl SiteEnergy<Point<Cartesian<2>>> for Wall {
+///     fn site_energy(&self, site_properties: &Point<Cartesian<2>>) -> f64 {
+///         if site_properties.position[1].abs() < 1.0 {
+///             f64::INFINITY
+///         } else {
+///             0.0
+///         }
+///     }
+///
+///     fn is_only_infinite_or_zero() -> bool {
+///         true
+///     }
+/// }
+///
+/// fn main() -> Result<(), Box<dyn std::error::Error>> {
+///     let mut microstate = Microstate::new();
+///     microstate.extend_bodies([
+///         Body::point(Cartesian::from([1.0, 1.25])),
+///         Body::point(Cartesian::from([-1.0, 2.0])),
+///     ])?;
+///
+///     let wall = External(Wall);
+///
+///     let total_energy = wall.total_energy(&microstate);
+///     assert_eq!(total_energy, 0.0);
+///     Ok(())
+/// }
+/// ```
 pub struct External<E>(pub E);
 
 impl<B, S, X, C, E> TotalEnergy<Microstate<B, S, X, C>> for External<E>
@@ -60,8 +94,9 @@ where
     /// evaluated only at the body centers. In general, hoomd-rs interactions apply
     /// to sites. Use a custom implementation to compute energies over body centers.
     ///
-    /// # Example
+    /// # Examples
     ///
+    /// A linear external potential:
     /// ```
     /// use hoomd_interaction::{External, TotalEnergy, external::Linear};
     /// use hoomd_microstate::{Body, Microstate, property::Point};
@@ -85,43 +120,58 @@ where
     /// # Ok(())
     /// # }
     /// ```
+    ///
+    /// Infinite interaction with a wall:
+    /// ```
+    /// use hoomd_interaction::{External, SiteEnergy, TotalEnergy};
+    /// use hoomd_microstate::{Body, Microstate, property::Point};
+    /// use hoomd_vector::Cartesian;
+    ///
+    /// struct Wall;
+    ///
+    /// impl SiteEnergy<Point<Cartesian<2>>> for Wall {
+    ///     fn site_energy(&self, site_properties: &Point<Cartesian<2>>) -> f64 {
+    ///         if site_properties.position[1].abs() < 1.0 {
+    ///             f64::INFINITY
+    ///         } else {
+    ///             0.0
+    ///         }
+    ///     }
+    ///
+    ///     fn is_only_infinite_or_zero() -> bool {
+    ///         true
+    ///     }
+    /// }
+    ///
+    /// fn main() -> Result<(), Box<dyn std::error::Error>> {
+    ///     let mut microstate = Microstate::new();
+    ///     microstate.extend_bodies([
+    ///         Body::point(Cartesian::from([1.0, 1.25])),
+    ///         Body::point(Cartesian::from([-1.0, 2.0])),
+    ///     ])?;
+    ///
+    ///     let wall = External(Wall);
+    ///
+    ///     let total_energy = wall.total_energy(&microstate);
+    ///     assert_eq!(total_energy, 0.0);
+    ///     Ok(())
+    /// }
+    /// ```
     #[inline]
     fn total_energy(&self, microstate: &Microstate<B, S, X, C>) -> f64 {
-        microstate
-            .sites()
-            .iter()
-            .fold(0.0, |total, s| total + self.0.site_energy(&s.properties))
+        let mut total = 0.0;
+        for site in microstate.sites() {
+            let one = self.0.site_energy(&site.properties);
+            if one == f64::INFINITY {
+                return one;
+            }
+            total += one;
+        }
+
+        total
     }
 }
 
-/// Evaluate the change in energy contributed by `External` when a single body is updated.
-///
-/// # Example
-///
-/// ```
-/// use hoomd_interaction::{DeltaEnergyOne, External, external::Linear};
-/// use hoomd_microstate::{Body, Microstate, property::Point};
-/// use hoomd_vector::Cartesian;
-///
-/// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-/// let mut microstate = Microstate::new();
-/// microstate.add_body(Body::point(Cartesian::from([0.0, 0.0])))?;
-///
-/// let linear = External(Linear {
-///     alpha: 1.0,
-///     plane_origin: Cartesian::default(),
-///     plane_normal: [0.0, 1.0].try_into()?,
-/// });
-///
-/// let delta_energy = linear.delta_energy_one(
-///     &microstate,
-///     0,
-///     &Body::point([0.0, -1.0].into()),
-/// );
-/// assert_eq!(delta_energy, -1.0);
-/// # Ok(())
-/// # }
-/// ```
 impl<P, B, S, X, C, E> DeltaEnergyOne<B, S, X, C> for External<E>
 where
     E: SiteEnergy<S>,
@@ -129,6 +179,76 @@ where
     S: Position<Position = P>,
     C: Wrap<B> + Wrap<S>,
 {
+    /// Evaluate the change in energy contributed by `External` when a single body is updated.
+    ///
+    /// # Examples
+    ///
+    /// A linear external potential:
+    /// ```
+    /// use hoomd_interaction::{DeltaEnergyOne, External, external::Linear};
+    /// use hoomd_microstate::{Body, Microstate, property::Point};
+    /// use hoomd_vector::Cartesian;
+    ///
+    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
+    /// let mut microstate = Microstate::new();
+    /// microstate.add_body(Body::point(Cartesian::from([0.0, 0.0])))?;
+    ///
+    /// let linear = External(Linear {
+    ///     alpha: 1.0,
+    ///     plane_origin: Cartesian::default(),
+    ///     plane_normal: [0.0, 1.0].try_into()?,
+    /// });
+    ///
+    /// let delta_energy = linear.delta_energy_one(
+    ///     &microstate,
+    ///     0,
+    ///     &Body::point([0.0, -1.0].into()),
+    /// );
+    /// assert_eq!(delta_energy, -1.0);
+    /// # Ok(())
+    /// # }
+    /// ```
+    ///
+    /// Infinite interaction with a wall:
+    /// ```
+    /// use hoomd_interaction::{DeltaEnergyOne, External, SiteEnergy};
+    /// use hoomd_microstate::{Body, Microstate, property::Point};
+    /// use hoomd_vector::Cartesian;
+    ///
+    /// struct Wall;
+    ///
+    /// impl SiteEnergy<Point<Cartesian<2>>> for Wall {
+    ///     fn site_energy(&self, site_properties: &Point<Cartesian<2>>) -> f64 {
+    ///         if site_properties.position[1].abs() < 1.0 {
+    ///             f64::INFINITY
+    ///         } else {
+    ///             0.0
+    ///         }
+    ///     }
+    ///
+    ///     fn is_only_infinite_or_zero() -> bool {
+    ///         true
+    ///     }
+    /// }
+    ///
+    /// fn main() -> Result<(), Box<dyn std::error::Error>> {
+    ///     let mut microstate = Microstate::new();
+    ///     microstate.extend_bodies([
+    ///         Body::point(Cartesian::from([1.0, 1.25])),
+    ///         Body::point(Cartesian::from([-1.0, 2.0])),
+    ///     ])?;
+    ///
+    ///     let wall = External(Wall);
+    ///
+    ///     let delta_energy = wall.delta_energy_one(
+    ///         &microstate,
+    ///         0,
+    ///         &Body::point([0.0, -0.5].into()),
+    ///     );
+    ///     assert_eq!(delta_energy, f64::INFINITY);
+    ///     Ok(())
+    /// }
+    /// ```
     #[inline]
     fn delta_energy_one(
         &self,
@@ -142,44 +262,31 @@ where
                 .boundary()
                 .wrap(final_body.properties.transform(s))
             {
-                Ok(wrapped_site) => energy_final += self.site_energy(&wrapped_site),
+                Ok(wrapped_site) => {
+                    let one = self.0.site_energy(&wrapped_site);
+                    if one == f64::INFINITY {
+                        return one;
+                    }
+                    energy_final += one;
+                }
                 Err(_) => return f64::INFINITY,
             }
         }
 
-        let energy_initial = initial_microstate
-            .iter_body_sites(body_index)
-            .fold(0.0, |total, s| total + self.site_energy(&s.properties));
+        let energy_initial = if E::is_only_infinite_or_zero() {
+            0.0
+        } else {
+            initial_microstate
+                .iter_body_sites(body_index)
+                .fold(0.0, |total, s| {
+                    total + self.0.site_energy_initial(&s.properties)
+                })
+        };
 
         energy_final - energy_initial
     }
 }
 
-/// Evaluate the change in energy contributed by `External` when a single body is inserted.
-///
-/// # Example
-///
-/// ```
-/// use hoomd_interaction::{DeltaEnergyInsert, External, external::Linear};
-/// use hoomd_microstate::{Body, Microstate, property::Point};
-/// use hoomd_vector::Cartesian;
-///
-/// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-/// let mut microstate = Microstate::new();
-/// microstate.add_body(Body::point(Cartesian::from([0.0, 0.0])))?;
-///
-/// let linear = External(Linear {
-///     alpha: 1.0,
-///     plane_origin: Cartesian::default(),
-///     plane_normal: [0.0, 1.0].try_into()?,
-/// });
-///
-/// let delta_energy = linear
-///     .delta_energy_insert(&microstate, &Body::point([0.0, -1.0].into()));
-/// assert_eq!(delta_energy, -1.0);
-/// # Ok(())
-/// # }
-/// ```
 impl<P, B, S, X, C, E> DeltaEnergyInsert<B, S, X, C> for External<E>
 where
     E: SiteEnergy<S>,
@@ -187,6 +294,70 @@ where
     S: Position<Position = P>,
     C: Wrap<B> + Wrap<S>,
 {
+    /// Evaluate the change in energy contributed by `External` when a single body is inserted.
+    ///
+    /// # Examples
+    ///
+    /// A linear external potential:
+    /// ```
+    /// use hoomd_interaction::{DeltaEnergyInsert, External, external::Linear};
+    /// use hoomd_microstate::{Body, Microstate, property::Point};
+    /// use hoomd_vector::Cartesian;
+    ///
+    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
+    /// let mut microstate = Microstate::new();
+    /// microstate.add_body(Body::point(Cartesian::from([0.0, 0.0])))?;
+    ///
+    /// let linear = External(Linear {
+    ///     alpha: 1.0,
+    ///     plane_origin: Cartesian::default(),
+    ///     plane_normal: [0.0, 1.0].try_into()?,
+    /// });
+    ///
+    /// let delta_energy = linear
+    ///     .delta_energy_insert(&microstate, &Body::point([0.0, -1.0].into()));
+    /// assert_eq!(delta_energy, -1.0);
+    /// # Ok(())
+    /// # }
+    /// ```
+    ///
+    /// Infinite interaction with a wall:
+    /// ```
+    /// use hoomd_interaction::{DeltaEnergyInsert, External, SiteEnergy};
+    /// use hoomd_microstate::{Body, Microstate, property::Point};
+    /// use hoomd_vector::Cartesian;
+    ///
+    /// struct Wall;
+    ///
+    /// impl SiteEnergy<Point<Cartesian<2>>> for Wall {
+    ///     fn site_energy(&self, site_properties: &Point<Cartesian<2>>) -> f64 {
+    ///         if site_properties.position[1].abs() < 1.0 {
+    ///             f64::INFINITY
+    ///         } else {
+    ///             0.0
+    ///         }
+    ///     }
+    ///
+    ///     fn is_only_infinite_or_zero() -> bool {
+    ///         true
+    ///     }
+    /// }
+    ///
+    /// fn main() -> Result<(), Box<dyn std::error::Error>> {
+    ///     let mut microstate = Microstate::new();
+    ///     microstate.extend_bodies([
+    ///         Body::point(Cartesian::from([1.0, 1.25])),
+    ///         Body::point(Cartesian::from([-1.0, 2.0])),
+    ///     ])?;
+    ///
+    ///     let wall = External(Wall);
+    ///
+    ///     let delta_energy = wall
+    ///         .delta_energy_insert(&microstate, &Body::point([0.0, -0.5].into()));
+    ///     assert_eq!(delta_energy, f64::INFINITY);
+    ///     Ok(())
+    /// }
+    /// ```
     #[inline]
     fn delta_energy_insert(
         &self,
@@ -199,7 +370,13 @@ where
                 .boundary()
                 .wrap(new_body.properties.transform(s))
             {
-                Ok(wrapped_site) => energy_final += self.site_energy(&wrapped_site),
+                Ok(wrapped_site) => {
+                    let one = self.0.site_energy(&wrapped_site);
+                    if one == f64::INFINITY {
+                        return one;
+                    }
+                    energy_final += one;
+                }
                 Err(_) => return f64::INFINITY,
             }
         }
@@ -208,62 +385,97 @@ where
     }
 }
 
-/// Evaluate the change in energy contributed by `External` when a single body is removed.
-///
-/// # Example
-///
-/// ```
-/// use hoomd_interaction::{DeltaEnergyRemove, External, external::Linear};
-/// use hoomd_microstate::{Body, Microstate, property::Point};
-/// use hoomd_vector::Cartesian;
-///
-/// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-/// let mut microstate = Microstate::new();
-/// microstate.add_body(Body::point(Cartesian::from([0.0, 1.0])))?;
-///
-/// let linear = External(Linear {
-///     alpha: 1.0,
-///     plane_origin: Cartesian::default(),
-///     plane_normal: [0.0, 1.0].try_into()?,
-/// });
-///
-/// let delta_energy = linear.delta_energy_remove(&microstate, 0);
-/// assert_eq!(delta_energy, -1.0);
-/// # Ok(())
-/// # }
-/// ```
 impl<B, S, X, C, E> DeltaEnergyRemove<B, S, X, C> for External<E>
 where
     E: SiteEnergy<S>,
 {
+    /// Evaluate the change in energy contributed by `External` when a single body is removed.
+    ///
+    /// # Examples
+    ///
+    /// A linear external potential:
+    /// ```
+    /// use hoomd_interaction::{DeltaEnergyRemove, External, external::Linear};
+    /// use hoomd_microstate::{Body, Microstate, property::Point};
+    /// use hoomd_vector::Cartesian;
+    ///
+    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
+    /// let mut microstate = Microstate::new();
+    /// microstate.add_body(Body::point(Cartesian::from([0.0, 1.0])))?;
+    ///
+    /// let linear = External(Linear {
+    ///     alpha: 1.0,
+    ///     plane_origin: Cartesian::default(),
+    ///     plane_normal: [0.0, 1.0].try_into()?,
+    /// });
+    ///
+    /// let delta_energy = linear.delta_energy_remove(&microstate, 0);
+    /// assert_eq!(delta_energy, -1.0);
+    /// # Ok(())
+    /// # }
+    /// ```
+    ///
+    /// Infinite interaction with a wall:
+    /// ```
+    /// use hoomd_interaction::{DeltaEnergyRemove, External, SiteEnergy};
+    /// use hoomd_microstate::{Body, Microstate, property::Point};
+    /// use hoomd_vector::Cartesian;
+    ///
+    /// struct Wall;
+    ///
+    /// impl SiteEnergy<Point<Cartesian<2>>> for Wall {
+    ///     fn site_energy(&self, site_properties: &Point<Cartesian<2>>) -> f64 {
+    ///         if site_properties.position[1].abs() < 1.0 {
+    ///             f64::INFINITY
+    ///         } else {
+    ///             0.0
+    ///         }
+    ///     }
+    ///
+    ///     fn is_only_infinite_or_zero() -> bool {
+    ///         true
+    ///     }
+    /// }
+    ///
+    /// fn main() -> Result<(), Box<dyn std::error::Error>> {
+    ///     let mut microstate = Microstate::new();
+    ///     microstate.extend_bodies([
+    ///         Body::point(Cartesian::from([1.0, 1.25])),
+    ///         Body::point(Cartesian::from([-1.0, 2.0])),
+    ///     ])?;
+    ///
+    ///     let wall = External(Wall);
+    ///
+    ///     let delta_energy = wall.delta_energy_remove(&microstate, 0);
+    ///     assert_eq!(delta_energy, 0.0);
+    ///     Ok(())
+    /// }
+    /// ```
     #[inline]
     fn delta_energy_remove(
         &self,
         initial_microstate: &Microstate<B, S, X, C>,
         body_index: usize,
     ) -> f64 {
+        if E::is_only_infinite_or_zero() {
+            return 0.0;
+        }
+
         let energy_initial = initial_microstate
             .iter_body_sites(body_index)
-            .fold(0.0, |total, s| total + self.site_energy(&s.properties));
+            .fold(0.0, |total, s| {
+                total + self.0.site_energy_initial(&s.properties)
+            });
 
         -energy_initial
     }
 }
 
-impl<E, S> SiteEnergy<S> for External<E>
-where
-    E: SiteEnergy<S>,
-{
-    #[inline]
-    fn site_energy(&self, site_properties: &S) -> f64 {
-        self.0.site_energy(site_properties)
-    }
-}
-
 #[cfg(test)]
-mod tests {
+mod test_finite {
     use super::*;
     use crate::external::Linear;
+    use assert2::check;
     use hoomd_geometry::shape::Rectangle;
     use hoomd_microstate::{
         Body, Microstate,
@@ -314,7 +526,7 @@ mod tests {
             let test_se = TestSE;
             let single = External(test_se);
 
-            assert_eq!(single.total_energy(&microstate), 3.0);
+            check!(single.total_energy(&microstate) == 3.0);
         }
 
         #[rstest]
@@ -329,8 +541,8 @@ mod tests {
             let test_se = TestSE;
             let single = External(test_se);
 
-            assert_eq!(single.site_energy(&microstate.sites()[0].properties), 1.0);
-            assert_eq!(single.site_energy(&microstate.sites()[1].properties), 2.0);
+            check!(single.0.site_energy(&microstate.sites()[0].properties) == 1.0);
+            check!(single.0.site_energy(&microstate.sites()[1].properties) == 2.0);
         }
     }
     mod delta_energy {
@@ -367,14 +579,8 @@ mod tests {
 
             let energy = External(Zero);
 
-            assert_eq!(
-                energy.delta_energy_one(&microstate, 0, &final_body),
-                f64::INFINITY
-            );
-            assert_eq!(
-                energy.delta_energy_insert(&microstate, &final_body),
-                f64::INFINITY
-            );
+            check!(energy.delta_energy_one(&microstate, 0, &final_body) == f64::INFINITY);
+            check!(energy.delta_energy_insert(&microstate, &final_body) == f64::INFINITY);
         }
 
         #[test]
@@ -406,9 +612,217 @@ mod tests {
                 alpha: 4.0,
             });
 
-            assert_eq!(energy.delta_energy_one(&microstate, 0, &final_body), 2.0);
-            assert_eq!(energy.delta_energy_insert(&microstate, &final_body), 6.0);
-            assert_eq!(energy.delta_energy_remove(&microstate, 0), -4.0);
+            check!(energy.delta_energy_one(&microstate, 0, &final_body) == 2.0);
+            check!(energy.delta_energy_insert(&microstate, &final_body) == 6.0);
+            check!(energy.delta_energy_remove(&microstate, 0) == -4.0);
+        }
+    }
+}
+
+#[cfg(test)]
+mod test_infinite {
+    use super::*;
+    use assert2::check;
+    use hoomd_geometry::shape::Rectangle;
+    use hoomd_microstate::{
+        Body, Microstate,
+        boundary::{Closed, Open},
+        property::{Point, Position},
+    };
+    use hoomd_vector::Cartesian;
+    use rstest::*;
+
+    struct TestSO;
+
+    impl<S> SiteEnergy<S> for TestSO
+    where
+        S: Position<Position = Cartesian<2>>,
+    {
+        fn site_energy(&self, site_properties: &S) -> f64 {
+            if site_properties.position()[1].abs() < 1.0 {
+                f64::INFINITY
+            } else {
+                0.0
+            }
+        }
+
+        fn is_only_infinite_or_zero() -> bool {
+            true
+        }
+    }
+
+    mod site_energy {
+        use super::*;
+        use hoomd_microstate::SiteKey;
+        use hoomd_spatial::AllPairs;
+
+        #[fixture]
+        fn microstate()
+        -> Microstate<Point<Cartesian<2>>, Point<Cartesian<2>>, AllPairs<SiteKey>, Open> {
+            let mut microstate = Microstate::new();
+            microstate
+                .extend_bodies([
+                    Body::point(Cartesian::from([1.0, -2.0])),
+                    Body::point(Cartesian::from([-1.0, 3.0])),
+                ])
+                .expect("hard-coded bodies should be in the boundary");
+            microstate
+        }
+
+        #[fixture]
+        fn overlapping_microstate()
+        -> Microstate<Point<Cartesian<2>>, Point<Cartesian<2>>, AllPairs<SiteKey>, Open> {
+            let mut microstate = Microstate::new();
+            microstate
+                .extend_bodies([
+                    Body::point(Cartesian::from([1.0, 0.75])),
+                    Body::point(Cartesian::from([-1.0, 3.0])),
+                ])
+                .expect("hard-coded bodies should be in the boundary");
+            microstate
+        }
+
+        #[rstest]
+        fn single_total_0(
+            microstate: Microstate<
+                Point<Cartesian<2>>,
+                Point<Cartesian<2>>,
+                AllPairs<SiteKey>,
+                Open,
+            >,
+        ) {
+            let single = External(TestSO);
+
+            check!(single.total_energy(&microstate) == 0.0);
+        }
+
+        #[rstest]
+        fn single_total_inf(
+            overlapping_microstate: Microstate<
+                Point<Cartesian<2>>,
+                Point<Cartesian<2>>,
+                AllPairs<SiteKey>,
+                Open,
+            >,
+        ) {
+            let single = External(TestSO);
+
+            check!(single.total_energy(&overlapping_microstate) == f64::INFINITY);
+        }
+
+        #[rstest]
+        fn single_site_0(
+            microstate: Microstate<
+                Point<Cartesian<2>>,
+                Point<Cartesian<2>>,
+                AllPairs<SiteKey>,
+                Open,
+            >,
+        ) {
+            let single = External(TestSO);
+
+            check!(single.0.site_energy(&microstate.sites()[0].properties) == 0.0);
+            check!(single.0.site_energy(&microstate.sites()[1].properties) == 0.0);
+        }
+
+        #[rstest]
+        fn single_site_inf(
+            overlapping_microstate: Microstate<
+                Point<Cartesian<2>>,
+                Point<Cartesian<2>>,
+                AllPairs<SiteKey>,
+                Open,
+            >,
+        ) {
+            let single = External(TestSO);
+
+            check!(
+                single
+                    .0
+                    .site_energy(&overlapping_microstate.sites()[0].properties)
+                    == f64::INFINITY
+            );
+            check!(
+                single
+                    .0
+                    .site_energy(&overlapping_microstate.sites()[1].properties)
+                    == 0.0
+            );
+        }
+    }
+    mod delta_energy {
+        use super::*;
+
+        struct Zero;
+
+        impl SiteEnergy<Point<Cartesian<2>>> for Zero {
+            fn site_energy(&self, _site_properties: &Point<Cartesian<2>>) -> f64 {
+                0.0
+            }
+
+            fn is_only_infinite_or_zero() -> bool {
+                true
+            }
+        }
+
+        #[test]
+        fn site_outside() {
+            let cuboid = Rectangle::with_equal_edges(
+                4.0.try_into()
+                    .expect("hard-coded constant should be positive"),
+            );
+            let square = Closed(cuboid);
+
+            let body = Body {
+                properties: Point::new(Cartesian::from([0.0, 0.0])),
+                sites: [Point::new(Cartesian::from([1.0, 0.0]))].into(),
+            };
+            let mut final_body = body.clone();
+            final_body.properties.position[0] = 1.0;
+
+            let microstate = Microstate::builder()
+                .boundary(square)
+                .bodies([body])
+                .try_build()
+                .expect("the hard-coded bodies should be in the boundary");
+
+            let energy = External(Zero);
+
+            check!(energy.delta_energy_one(&microstate, 0, &final_body) == f64::INFINITY);
+            check!(energy.delta_energy_insert(&microstate, &final_body) == f64::INFINITY);
+        }
+
+        #[test]
+        fn delta_energy() {
+            let cuboid = Rectangle::with_equal_edges(
+                4.0.try_into()
+                    .expect("hard-coded constant should be positive"),
+            );
+            let square = Closed(cuboid);
+
+            let body = Body {
+                properties: Point::new(Cartesian::from([1.5, 1.5])),
+                sites: [Point::new(Cartesian::from([0.0, 0.0]))].into(),
+            };
+            let mut final_body_inf = body.clone();
+            final_body_inf.properties.position[1] = 0.5;
+
+            let mut final_body_0 = body.clone();
+            final_body_0.properties.position[1] = -1.5;
+
+            let microstate = Microstate::builder()
+                .boundary(square)
+                .bodies([body])
+                .try_build()
+                .expect("the hard-coded bodies should be in the boundary");
+
+            let energy = External(TestSO);
+
+            check!(energy.delta_energy_one(&microstate, 0, &final_body_0) == 0.0);
+            check!(energy.delta_energy_one(&microstate, 0, &final_body_inf) == f64::INFINITY);
+            check!(energy.delta_energy_insert(&microstate, &final_body_0) == 0.0);
+            check!(energy.delta_energy_insert(&microstate, &final_body_inf) == f64::INFINITY);
+            check!(energy.delta_energy_remove(&microstate, 0) == 0.0);
         }
     }
 }
