+++ hoomd-interaction/src/lib.rs
@@ -19,14 +19,12 @@ pub mod pairwise;
 
 mod cutoff_pair;
 mod cutoff_pair_overlap;
-mod external_overlap;
 mod external_type;
 mod hamiltonian;
 mod zero;
 
 pub use cutoff_pair::CutoffPair;
 pub use cutoff_pair_overlap::CutoffPairOverlap;
-pub use external_overlap::ExternalOverlap;
 pub use external_type::External;
 pub use zero::Zero;
 
@@ -95,7 +93,7 @@ pub trait TotalEnergy<M> {
 ///
 /// ## Examples
 ///
-/// Implement a custom site energy method:
+/// Implement a custom site energy function:
 ///
 /// ```
 /// use hoomd_interaction::{External, SiteEnergy, TotalEnergy};
@@ -135,31 +133,11 @@ pub trait TotalEnergy<M> {
 /// # Ok(())
 /// # }
 /// ```
-pub trait SiteEnergy<S> {
-    /// Evaluate the energy contribution of a single site.
-    #[must_use]
-    fn site_energy(&self, site_properties: &S) -> f64;
-}
-
-/// Check if a site overlaps with an object external to the microstate.
 ///
-/// The `SiteOverlap` trait describes a type that can determine whether or not
-/// a site overlaps with another object *as a function only of that site's
-/// properties*.
-///
-/// The [`external`] module provides a number of commonly used implementations.
-/// Combine them with [`ExternalOverlap`] newtype for use with MC simulations or to
-/// compute system-wide properties.
-///
-/// The generic type names are:
-/// * `S`: The [`Site::properties`](hoomd_microstate::Site) type.
-///
-/// ## Examples
-///
-/// Implement a custom site overlap method:
+/// Custom method that checks for overlaps of a disk with a circular boundary.
 ///
 /// ```
-/// use hoomd_interaction::{ExternalOverlap, SiteOverlap, TotalEnergy};
+/// use hoomd_interaction::{External, SiteEnergy, TotalEnergy};
 /// use hoomd_microstate::{
 ///     Body, Microstate,
 ///     property::{Point, Position},
@@ -170,14 +148,22 @@ pub trait SiteEnergy<S> {
 ///     r: f64,
 /// }
 ///
-/// impl<S> SiteOverlap<S> for Custom
+/// impl<S> SiteEnergy<S> for Custom
 /// where
 ///     S: Position<Position = Cartesian<2>>,
 /// {
-///     /// Check for overlaps of a disk with a circular boundary.
-///     fn site_overlap(&self, site_properties: &S) -> bool {
-///         site_properties.position().distance(&Cartesian::default())
+///     fn site_energy(&self, site_properties: &S) -> f64 {
+///         if site_properties.position().distance(&Cartesian::default())
 ///             > self.r - 0.5
+///         {
+///             f64::INFINITY
+///         } else {
+///             0.0
+///         }
+///     }
+///
+///     fn is_only_infinite_or_zero() -> bool {
+///         true
 ///     }
 /// }
 ///
@@ -186,20 +172,49 @@ pub trait SiteEnergy<S> {
 /// microstate.extend_bodies([Body::point(Cartesian::from([9.6, 0.0]))])?;
 ///
 /// let custom_evaluator = Custom { r: 10.0 };
-/// let site_overlap =
-///     custom_evaluator.site_overlap(&microstate.sites()[0].properties);
-/// assert!(site_overlap);
+/// let site_energy =
+///     custom_evaluator.site_energy(&microstate.sites()[0].properties);
+/// assert_eq!(site_energy, f64::INFINITY);
 ///
-/// let custom = ExternalOverlap(custom_evaluator);
+/// let custom = External(custom_evaluator);
 /// let total_energy = custom.total_energy(&microstate);
 /// assert_eq!(total_energy, f64::INFINITY);
 /// # Ok(())
 /// # }
 /// ```
-pub trait SiteOverlap<S> {
-    /// Determine if a site overlaps with an object external to the microstate.
+pub trait SiteEnergy<S> {
+    /// Evaluate the energy contribution of a single site.
     #[must_use]
-    fn site_overlap(&self, site_properties: &S) -> bool;
+    fn site_energy(&self, site_properties: &S) -> f64;
+
+    /// Evaluate the energy contribution of a single site *in the initial state*.
+    ///
+    /// Override this method in potentials that have both infinite or zero
+    /// terms and finite terms, such as the sum of a hard site-wall interaction
+    /// plus an attractive well. `site_energy` should compute both terms and
+    /// `site_energy_initial` should compute only the finite terms.
+    ///
+    /// [`External`] calls `site_energy_initial` when evaluating the energy of
+    /// the initial state in a trial move. The infinite interaction term can be
+    /// assumed 0 in the initial state because no site will ever be placed in an
+    /// infinite energy configuration.
+    #[must_use]
+    #[inline]
+    fn site_energy_initial(&self, site_properties: &S) -> f64 {
+        self.site_energy(site_properties)
+    }
+
+    /// Does this potential only ever return infinity or zero?
+    ///
+    /// Override this method and return `true` for e.g. hard site-wall
+    /// interactions that always return infinity or zero and **never** any other
+    /// value. When this method returns `true`, [`External`] skips the initial
+    /// energy computation and assumes it is zero.
+    #[must_use]
+    #[inline]
+    fn is_only_infinite_or_zero() -> bool {
+        false
+    }
 }
 
 /// Compute the energy contribution from a pair of sites.
